Move color difference functions to own file
